% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_get.R
\name{cf_get}
\alias{cf_get}
\title{Download CalFlora metadata}
\usage{
cf_get(what_to_get, api_key = Sys.getenv("CALFLORA_API_KEY"))
}
\arguments{
\item{what_to_get}{A character string specifying which CalFlora API
endpoint to query. See the CalFlora API documentation for available
endpoints. For example, use \code{"columnsets"} to retrieve column-set
metadata used by the observations endpoint.}

\item{api_key}{Your CalFlora API key. By default, the function looks for
an API key stored in the \code{CALFLORA_API_KEY} environment variable.}
}
\value{
A data frame containing metadata returned by the CalFlora API.
}
\description{
Download basic metadata for different types of CalFlora data.
}
\details{
This function is useful for retrieving information about available
values for query parameters used when downloading CalFlora data.
}
\examples{
\dontrun{
calflora_columnsets <- cf_get("columnsets")

calflora_projects <- cf_get("projects")
}

}
