% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_observations_df.R
\name{cf_observations_df}
\alias{cf_observations_df}
\title{Download CalFlora observations}
\usage{
cf_observations_df(params, api_key = Sys.getenv("CALFLORA_API_KEY"))
}
\arguments{
\item{params}{A named list of query parameters to pass to the CalFlora
observations API. See the examples for commonly used parameters.}

\item{api_key}{Your CalFlora API key. By default, the function looks for
an API key stored in the \code{CALFLORA_API_KEY} environment variable.}
}
\value{
A data frame containing observation records returned by the
CalFlora API.
}
\description{
Download observation records from the CalFlora API.
}
\details{
The CalFlora observations API limits individual requests to 2,000
records. To download more than 2,000 records, use additional query
parameters to divide the request into smaller chunks and combine the
resulting data frames.
}
\examples{
\dontrun{
my_params <- list(
  shapeId = "rs1515",
  dateAfter = "1960-01-01",
  dateBefore = "2025-12-31",
  csetId = "ACR-WT",
  projectIds = "pr940"
)

observations <- cf_observations_df(params = my_params)
}

}
